refactor: Sliding/NonSliding 구분 제거 및 패키지 이동
