+++ src/main/java/chess/piece/PiecesGenrator.java
@@ -1,5 +1,6 @@
 package chess.piece;
 
+import chess.piece.type.Piece;
 import java.util.List;
 
 public interface PiecesGenrator {
