--- src/main/java/chess/piece/sliding/SlidingPiece.java
@@ -1,33 +1,0 @@
-package chess.piece.sliding;
-
-import chess.board.Position;
-import chess.piece.Direction;
-import chess.piece.Piece;
-import chess.piece.Side;
-import java.util.ArrayList;
-import java.util.List;
-
-public abstract class SlidingPiece extends Piece {
-
-    public SlidingPiece(final Position position, final Side side) {
-        super(position, side);
-    }
-
-    @Override
-    public List<Position> getPaths(Position targetPosition) {
-        List<Position> paths = new ArrayList<>();
-        final Direction direction = position.getDirectionTo(targetPosition);
-        final int moveCountBeforeArrivalPosition = position.getMoveCount(targetPosition, direction) - 1;
-        Position nextPosition = this.position;
-        for (int next = 0; next < moveCountBeforeArrivalPosition; next++) {
-            nextPosition = nextPosition.getNextPosition(direction);
-            paths.add(nextPosition);
-        }
-        return paths;
-    }
-
-    @Override
-    public boolean canPassThrough() {
-        return false;
-    }
-}
