+++ src/main/java/chess/piece/type/Pawn.java
@@ -1,6 +1,8 @@
-package chess.piece;
+package chess.piece.type;
 
 import chess.board.Position;
+import chess.piece.Direction;
+import chess.piece.Side;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
