+++ src/main/java/chess/piece/type/Queen.java
@@ -1,4 +1,4 @@
-package chess.piece.sliding;
+package chess.piece.type;
 
 import chess.board.Position;
 import chess.piece.Direction;
@@ -6,7 +6,7 @@ import chess.piece.Side;
 import java.util.ArrayList;
 import java.util.List;
 
-public class Queen extends SlidingPiece {
+public class Queen extends Piece {
 
     private final List<Direction> directions;
     public Queen(final Position position, final Side side) {
@@ -34,8 +34,25 @@ public class Queen extends SlidingPiece {
     }
 
     @Override
+    public List<Position> getPaths(Position targetPosition) {
+        List<Position> paths = new ArrayList<>();
+        final Direction direction = position.getDirectionTo(targetPosition);
+        final int moveCountBeforeArrivalPosition = position.getMoveCount(targetPosition, direction) - 1;
+        Position nextPosition = this.position;
+        for (int next = 0; next < moveCountBeforeArrivalPosition; next++) {
+            nextPosition = nextPosition.getNextPosition(direction);
+            paths.add(nextPosition);
+        }
+        return paths;
+    }
+
+    @Override
     public Queen move(final Position positionToMove) {
         return new Queen(positionToMove, this.side);
     }
 
+    @Override
+    public boolean canPassThrough() {
+        return false;
+    }
 }
