+++ src/main/java/chess/piece/type/Rook.java
@@ -1,4 +1,4 @@
-package chess.piece.sliding;
+package chess.piece.type;
 
 import chess.board.Position;
 import chess.piece.Direction;
@@ -6,7 +6,7 @@ import chess.piece.Side;
 import java.util.ArrayList;
 import java.util.List;
 
-public class Rook extends SlidingPiece {
+public class Rook extends Piece {
 
     private final List<Direction> directions;
 
@@ -31,8 +31,25 @@ public class Rook extends SlidingPiece {
     }
 
     @Override
+    public List<Position> getPaths(Position targetPosition) {
+        List<Position> paths = new ArrayList<>();
+        final Direction direction = position.getDirectionTo(targetPosition);
+        final int moveCountBeforeArrivalPosition = position.getMoveCount(targetPosition, direction) - 1;
+        Position nextPosition = this.position;
+        for (int next = 0; next < moveCountBeforeArrivalPosition; next++) {
+            nextPosition = nextPosition.getNextPosition(direction);
+            paths.add(nextPosition);
+        }
+        return paths;
+    }
+
+    @Override
     public Rook move(final Position positionToMove) {
         return new Rook(positionToMove, this.side);
     }
 
+    @Override
+    public boolean canPassThrough() {
+        return false;
+    }
 }
