+++ src/main/java/chess/view/PieceName.java
@@ -1,12 +1,12 @@
 package chess.view;
 
-import chess.piece.sliding.Bishop;
-import chess.piece.nonsliding.King;
-import chess.piece.nonsliding.Knight;
-import chess.piece.Pawn;
-import chess.piece.Piece;
-import chess.piece.sliding.Queen;
-import chess.piece.sliding.Rook;
+import chess.piece.type.Bishop;
+import chess.piece.type.King;
+import chess.piece.type.Knight;
+import chess.piece.type.Pawn;
+import chess.piece.type.Piece;
+import chess.piece.type.Queen;
+import chess.piece.type.Rook;
 import chess.piece.Side;
 import java.util.Arrays;
