+++ src/test/java/chess/PiecesTest.java
@@ -7,8 +7,8 @@ import chess.board.File;
 import chess.board.Position;
 import chess.board.Rank;
 import chess.piece.AllPiecesGenerator;
-import chess.piece.Pawn;
-import chess.piece.Piece;
+import chess.piece.type.Pawn;
+import chess.piece.type.Piece;
 import chess.piece.Pieces;
 import chess.piece.Side;
 import java.util.List;
